norm exec
